requirements.txt
RPi.GPIO
spidev
Pillow
numpy
lgpio
requests
python-dotenv
evdev
qrcode[pil]
